backend/pyproject.toml
[project]
name = "bb-mybudgetbook-backend"
version = "0.1.0"
description = "Backend scaffold for BB-myBudgetBook"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.0,<2.0.0",
    "argon2-cffi>=23.1.0,<24.0.0",
    "email-validator>=2.2.0,<3.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "itsdangerous>=2.2.0,<3.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "pydantic-settings>=2.9.0,<3.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0,<1.0.0",
    "pytest>=8.3.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
